backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
numpy==1.24.4
opencv-python==4.8.1.78
pillow==10.3.0
pydantic==1.10.15
onnxruntime==1.17.3
python-multipart==0.0.9
